+++ hledger-lib/Hledger/Reports.hs
@@ -15,7 +15,6 @@ module Hledger.Reports (
   module Hledger.Reports.ReportTypes,
   module Hledger.Reports.EntriesReport,
   module Hledger.Reports.PostingsReport,
-  module Hledger.Reports.TransactionsReport,
   module Hledger.Reports.AccountTransactionsReport,
   module Hledger.Reports.BalanceReport,
   module Hledger.Reports.MultiBalanceReport,
@@ -30,7 +29,6 @@ import Hledger.Reports.ReportTypes
 import Hledger.Reports.AccountTransactionsReport
 import Hledger.Reports.EntriesReport
 import Hledger.Reports.PostingsReport
-import Hledger.Reports.TransactionsReport
 import Hledger.Reports.BalanceReport
 import Hledger.Reports.MultiBalanceReport
 import Hledger.Reports.BudgetReport
